+++ example/.che/project.json
@@ -1,0 +1,1 @@
+{"type":"com.sap.hcp.html5","builders":{"default":"npm-builder","configs":{}},"runners":{"configs":{}},"attributes":{"sap.watt.common.setting":["{\"translation\":{\"translationDomain\":\"\",\"supportedLanguages\":\"\",\"defaultLanguage\":\"\",\"defaultI18NPropertyFile\":\"\",\"resourceModelName\":\"\"},\"projectType\":[\"sap.watt.uitools.ide.fiori\",\"sap.watt.uitools.ide.web\",\"sap.watt.uitools.ide.feature\",\"com.sap.hcp.html5\"]}"]},"description":"","mixinTypes":["mtamodule"]}
