+++ example/.npmrc
@@ -1,0 +1,2 @@
+# Use SAP registry for @sap packages
+@sap:registry=https://npm.sap.com
