--- example/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<body>
-<script src="node_modules/@sap-webide/webide-client-tools/resources/local_env/webide_starter.js"></script>
-
-<script>
-    startWebIDE()
-</script>
-
-</body>
-</html>
