+++ example/package.json
@@ -1,26 +1,26 @@
 {
-    "name": "client-tools-example",
-    "version": "6.6.6",
-    "description": "Example use cases for client tools package - used as both Living Docs and Integration tests",
+    "name": "my_feature",
+    "title": "My Feature",
+    "version": "0.0.1",
+    "description": "Example feature",
     "scripts": {
         "test": "npm run bundle && node ./scripts/karma_ci --coverage --DIST",
         "dev_server": "node ./scripts/dev_server.js",
-        "bundle": "node ./scripts/bundle.js"
+        "bundle": "node ./scripts/bundle.js",
+        "lint": "eslint . --ext=.js,.json",
+        "postinstall": "npm run bundle && npm run lint"
     },
-    "bundledPlugins": {
-        "sap.webide.example.plugin": "file:src/webide-plugin-example"
-    },
-    "webidePreloads": {
-        "js": [
-            "config-preload.js",
-            "i18n/config-preload.js"
-        ],
-        "config": [
-            "config-preload.json"
-        ]
-    },
+    "files": [
+        "i18n",
+        "src",
+        "config-preload.js",
+        "config-preload.json",
+        "package.json"
+    ],
     "devDependencies": {
+        "@sap/eslint-plugin-webide-feature": "1.3.7",
         "chai": "^3.5.0",
+        "eslint": "4.12.1",
         "karma": "^1.7.0",
         "karma-chrome-launcher": "^2.1.1",
         "karma-coverage": "^1.1.1",
@@ -31,7 +31,19 @@
         "karma-requirejs": "^1.1.0",
         "minimist": "^1.2.0",
         "mocha": "3.4.2",
-        "webpack": "^4.2.0"
+        "webpack": "^4.2.0",
+        "@sap-webide/webide-client-tools": "latest"
     },
-    "dependencies": {}
+    "bundledPlugins": {
+        "my_plugin": "file:src/my_plugin"
+    },
+    "webidePreloads": {
+        "js": [
+            "config-preload.js",
+            "i18n/config-preload.js"
+        ],
+        "config": [
+            "config-preload.json"
+        ]
+    }
 }
