--- example/scripts/dev_server.js
@@ -1,4 +1,0 @@
-const clientTools = require("@sap-webide/webide-client-tools")
-
-// <startConnect> has an optional options argument which can be used to customize for different scenarios
-clientTools.devServer.startConnect()
