+++ example/src/my_plugin/i18n/i18n.properties
@@ -1,4 +1,4 @@
-#__ldi.translation.uuid=d60aef95-eae8-a199-865e-9cd9e035ffec
+#__ldi.translation.uuid=2ef20c8b-0532-af30-1327-5d50caa44a54
 # XTXT:
 sample_helloMessage = Hello {0}!
