--- example/src/webide-plugin-example/service/helloUtils.js
@@ -1,12 +1,0 @@
-/**
- * Some pure JS utilities for our sayHello service
- */
-define({
-    /**
-     * @param {string} msg
-     * @return {string}
-     */
-    addAngryMode: function(msg) {
-        return msg + "!??!!?"
-    }
-})
