+++ package.json
@@ -57,7 +57,7 @@
     "test_integration": "npm-run-all init_example execute_integration_tests unlink_example",
     "init_example": "node scripts/init_example.js",
     "execute_integration_tests": "jest --maxWorkers 1 --config test_integration/jest.config.json",
-    "unlink_example" : "cd ./example && npm unlink @sap-webide/webide-client-tools",
+    "unlink_example": "cd ./example && npm unlink @sap-webide/webide-client-tools",
     "docs": "shx rm -rf docs/web/html_docs && typedoc --tsconfig tsdocsconfig.json --out docs/web/html_docs --module commonjs --name WebideClientTools --excludeExternals --includeDeclarations",
     "format": "prettier --write \"**/*.js\"",
     "format_validate": "prettier --list-different \"**/*.js\"",
